feat: Rename segments in the provided folder
This commit updates the `audioDivisor.py` file to include a new function `rename_segments` that renames the segments in the provided folder. The function uses a regular expression pattern to match the filenames of the segments and then renames them by adding leading zeros to the number part of the filename. The function is used in the provided code to rename the segments in the "./data" folder.